requirements.txt
Django==1.4.4
South==0.7.6
distribute==0.6.31
dj-database-url==0.2.1
django-tastypie==0.9.12
gunicorn==0.16.1
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
pytz==2012j
six==1.2.0
wsgiref==0.1.2
